requirements.txt
numpy
scipy
matplotlib
ase
PyYAML
mpmath
